refact(ContactUs): hide it on mobile
